Fixed Simulator Opening
For some reason `open -a Simulator` stopped working and only opening by a full path works.

I was able to fix it manually by launching Xcode manually and closing it. Seems it does something more than `xcodebuild -runFirstLaunch`.

In any case, this change intorcudces a warmup script for Xcodes which opens all of them, waits a bit and verifies `open -a Simulator` works.